Refactor component structure and adjust import paths
The components related to 'Content' have been moved into a dedicated 'content' directory. All references and import paths in the relevant files have been adjusted to reflect this directory structure change. This reorganization makes the codebase easier to navigate and manage.
