+++ frontend/src/components/content/ContentResult/ContentResultHeader.component.test.tsx
