+++ frontend/src/components/content/ContentTypeIcon/ContentTypeIcon.tsx
@@ -1,6 +1,6 @@
 import { Box, Icon, Tooltip } from "@chakra-ui/react";
 import { FaFilm, FaTv } from "react-icons/fa6";
-import { contentTypeEnum } from "../../model/contentModel.ts";
+import { contentTypeEnum } from "../../../model/contentModel.ts";
 import { FaPaintBrush } from "react-icons/fa";
 
 type ContentCardTypeIconProps = {
