+++ frontend/src/pages/content/ContentPage.tsx
@@ -8,10 +8,10 @@ import {
 import { Container, Flex, Skeleton } from "@chakra-ui/react";
 import { useEffect, useState } from "react";
 import { getContent } from "../../services/contentService.ts";
-import ContentResultHeader from "../../components/ContentResult/ContentResultHeader.tsx";
+import ContentResultHeader from "../../components/content/ContentResult/ContentResultHeader.tsx";
 import NoData from "../../components/NoData/NoData.tsx";
-import ContentResultBody from "../../components/ContentResult/ContentResultBody.tsx";
-import ContentCard from "../../components/ContentCard/ContentCard.tsx";
+import ContentResultBody from "../../components/content/ContentResult/ContentResultBody.tsx";
+import ContentCard from "../../components/content/ContentCard/ContentCard.tsx";
 
 type ContentPageProps = {
   user: githubUserType | null | undefined;
